chore: reuse same class for faking credentials
