+++ pipeline-maven/src/test/java/org/jenkinsci/plugins/pipeline/maven/console/MaskPasswordsConsoleLogFilterTest.java
@@ -1,17 +1,14 @@
 package org.jenkinsci.plugins.pipeline.maven.console;
 
-import static com.cloudbees.plugins.credentials.CredentialsScope.GLOBAL;
-import static java.util.Arrays.asList;
-
 import java.util.ArrayList;
 import java.util.List;
 
-import org.acegisecurity.Authentication;
 import org.jenkinsci.plugins.configfiles.GlobalConfigFiles;
 import org.jenkinsci.plugins.configfiles.maven.GlobalMavenSettingsConfig;
 import org.jenkinsci.plugins.configfiles.maven.job.MvnGlobalSettingsProvider;
 import org.jenkinsci.plugins.configfiles.maven.security.ServerCredentialMapping;
 import org.jenkinsci.plugins.pipeline.maven.AbstractIntegrationTest;
+import org.jenkinsci.plugins.pipeline.maven.util.FakeCredentialsProvider;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
@@ -19,13 +16,9 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.jvnet.hudson.test.Issue;
 
-import com.cloudbees.plugins.credentials.Credentials;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
-import com.cloudbees.plugins.credentials.domains.DomainRequirement;
-import com.cloudbees.plugins.credentials.impl.UsernamePasswordCredentialsImpl;
 
 import hudson.ExtensionList;
-import hudson.model.ItemGroup;
 import hudson.model.Result;
 import jenkins.model.Jenkins;
 import jenkins.mvn.GlobalMavenConfig;
@@ -47,6 +40,7 @@ public class MaskPasswordsConsoleLogFilterTest extends AbstractIntegrationTest {
             "    }\n" +
             "}";
 
+        String serverId = "server-id";
         //@formatter:off
         String mavenGlobalSettings = "<?xml version='1.0' encoding='UTF-8'?>\n" +
             "<settings \n" +
@@ -55,17 +49,18 @@ public class MaskPasswordsConsoleLogFilterTest extends AbstractIntegrationTest {
             "        xsi:schemaLocation='http://maven.apache.org/SETTINGS/1.0.0 http://maven.apache.org/xsd/settings-1.0.0.xsd'>\n" +
             "    <servers>\n" +
             "       <server>\n" +
-            "           <id>server-id</id>\n" +
+            "           <id>" + serverId + "</id>\n" +
             "       </server>\n" +
             "    </servers>\n" +
             "</settings>\n";
         //@formatter:on
 
+        String credentialsId = "creds-id";
         ExtensionList<CredentialsProvider> extensionList = Jenkins.getInstance().getExtensionList(CredentialsProvider.class);
-        extensionList.add(extensionList.size(), new FakeCredentialsProvider(usernameIsSecret));
+        extensionList.add(extensionList.size(), new FakeCredentialsProvider(credentialsId, "aUser", "aPass", usernameIsSecret));
 
         List<ServerCredentialMapping> mappings = new ArrayList<>();
-        mappings.add(new ServerCredentialMapping("server-id", "creds-id"));
+        mappings.add(new ServerCredentialMapping(serverId, credentialsId));
         GlobalMavenSettingsConfig mavenGlobalSettingsConfig = new GlobalMavenSettingsConfig("maven-global-config-test", "maven-global-config-test", "",
                 mavenGlobalSettings, true, mappings);
 
@@ -77,9 +72,10 @@ public class MaskPasswordsConsoleLogFilterTest extends AbstractIntegrationTest {
             pipeline.setDefinition(new CpsFlowDefinition(pipelineScript, true));
             WorkflowRun build = jenkinsRule.assertBuildStatus(Result.SUCCESS, pipeline.scheduleBuild2(0));
             jenkinsRule.assertLogContains(
-                    "[withMaven] using Maven global settings.xml 'maven-global-config-test' with Maven servers credentials provided by Jenkins (replaceAll: true): [mavenServerId: 'server-id', jenkinsCredentials: 'creds-id']",
+                    "[withMaven] using Maven global settings.xml 'maven-global-config-test' with Maven servers credentials provided by Jenkins (replaceAll: true): [mavenServerId: '"
+                            + serverId + "', jenkinsCredentials: '" + credentialsId + "']",
                     build);
-            jenkinsRule.assertLogContains("<id>server-id</id>", build);
+            jenkinsRule.assertLogContains("<id>" + serverId + "</id>", build);
             if (usernameIsSecret) {
                 jenkinsRule.assertLogNotContains("<username>aUser</username>", build);
             } else {
@@ -88,32 +84,6 @@ public class MaskPasswordsConsoleLogFilterTest extends AbstractIntegrationTest {
             jenkinsRule.assertLogNotContains("<password>aPass</password>", build);
         } finally {
             GlobalMavenConfig.get().setSettingsProvider(null);
-        }
-    }
-
-    private static class FakeCredentialsProvider extends CredentialsProvider {
-        private boolean usernameIsSecret;
-
-        public FakeCredentialsProvider(boolean usernameIsSecret) {
-            this.usernameIsSecret = usernameIsSecret;
-        }
-
-        @Override
-        public boolean isEnabled(Object context) {
-            return true;
-        }
-
-        @Override
-        public <C extends Credentials> List<C> getCredentials(Class<C> type, ItemGroup itemGroup, Authentication authentication,
-                List<DomainRequirement> domainRequirements) {
-            UsernamePasswordCredentialsImpl creds = new UsernamePasswordCredentialsImpl(GLOBAL, "creds-id", "", "aUser", "aPass");
-            creds.setUsernameSecret(usernameIsSecret);
-            return (List<C>) asList(creds);
-        }
-
-        @Override
-        public <C extends Credentials> List<C> getCredentials(Class<C> type, ItemGroup itemGroup, Authentication authentication) {
-            return getCredentials(type, itemGroup, authentication, null);
         }
     }
