+++ pipeline-maven/src/test/java/org/jenkinsci/plugins/pipeline/maven/util/FakeCredentialsProvider.java
@@ -1,0 +1,49 @@
+package org.jenkinsci.plugins.pipeline.maven.util;
+
+import static com.cloudbees.plugins.credentials.CredentialsScope.GLOBAL;
+import static java.util.Arrays.asList;
+
+import java.util.List;
+
+import org.acegisecurity.Authentication;
+
+import com.cloudbees.plugins.credentials.Credentials;
+import com.cloudbees.plugins.credentials.CredentialsProvider;
+import com.cloudbees.plugins.credentials.domains.DomainRequirement;
+import com.cloudbees.plugins.credentials.impl.UsernamePasswordCredentialsImpl;
+
+import hudson.model.ItemGroup;
+
+public class FakeCredentialsProvider extends CredentialsProvider {
+
+    private String id;
+    private String username;
+    private String password;
+    private boolean usernameIsSecret;
+
+    public FakeCredentialsProvider(String id, String username, String password, boolean usernameIsSecret) {
+        this.id = id;
+        this.username = username;
+        this.password = password;
+        this.usernameIsSecret = usernameIsSecret;
+    }
+
+    @Override
+    public boolean isEnabled(Object context) {
+        return true;
+    }
+
+    @Override
+    public <C extends Credentials> List<C> getCredentials(Class<C> type, ItemGroup itemGroup, Authentication authentication,
+            List<DomainRequirement> domainRequirements) {
+        UsernamePasswordCredentialsImpl creds = new UsernamePasswordCredentialsImpl(GLOBAL, id, "", username, password);
+        creds.setUsernameSecret(usernameIsSecret);
+        return (List<C>) asList(creds);
+    }
+
+    @Override
+    public <C extends Credentials> List<C> getCredentials(Class<C> type, ItemGroup itemGroup, Authentication authentication) {
+        return getCredentials(type, itemGroup, authentication, null);
+    }
+
+}
